+++ kurtosis-module/impl/participant_network/cl/cl_client_launcher.go
@@ -14,7 +14,8 @@ type CLClientLauncher interface {
 		enclaveCtx *enclaves.EnclaveContext,
 		serviceId services.ServiceID,
 		image string,
-		logLevel module_io.ParticipantLogLevel,
+		participantLogLevel string,
+		globalLogLevel module_io.GlobalClientLogLevel,
 		// If nil, the node will be launched as a bootnode
 		bootnodeContext *CLClientContext,
 		elClientContext *el.ELClientContext,